feat: rename useCommunity file to hooks
